server/requirements.txt
haystack-ai
firebase-admin
python-dotenv
ollama
numpy
google-generativeai 